requirements.txt
Django==1.7.2
PyJWT==0.3.2
dj-database-url==0.3.0
django-toolbelt==0.0.1
djangorestframework==3.0.2
djangorestframework-jwt==1.1.1
gunicorn==19.1.1
jsonfield==1.0.0
psycopg2==2.5.4
whitenoise==1.0.6
wsgiref==0.1.2
